+++ src/controllers/GuideController.php
@@ -68,22 +68,24 @@ class GuideController extends Controller
     {
         $params = Craft::$app->getRequest()->getBodyParams();
 
-        $title = !empty($params['title']) ? $params['title'] : 'Guide saved on ' . Date('F jS \a\t g:ia');
-
         $slug = ($params['slug'] ?? false) ? $params['slug'] : $this->generateSlugFromTitle($title);
+        $title = !empty($params['title']) ? $params['title'] : 'Untitled Guide saved on ' . Date('F jS \a\t g:ia');
+        $totalGuides = Guide::$plugin->guide->getGuides([], 'count') ?? 0;
+        $weight = $params['weight'] ?? ($totalGuides + 1);
 
         $guide = new GuideModel([
             'authorId' => $params['authorId'],
             'content' => $params['contentEditor'],
             'contentCss' => $params['contentCss'],
             'contentJavascript' => $params['contentJavascript'],
-            'contentSource' => $params['contentSource'] ?? Guide::$pro ? 'field' : 'template',
+            'contentSource' => !empty($params['contentSource']) && Guide::$pro ? $params['contentSource'] : 'template',
             'contentUrl' => $params['contentUrl'],
             'renderMarkdown' => $params['renderMarkdown'],
             'slug' => $slug,
             'summary' => $params['summary'],
             'template' => $params['template'],
             'title' => $title,
+            'weight' => $weight,
         ]);
 
         if ($guide->validate()) {
@@ -91,8 +93,6 @@ class GuideController extends Controller
 
             return $this->redirect(UrlHelper::url($params['redirect'] ?? 'guide/list'));
         } else {
-            // @TODO handle error correctly
-//            Craft::dd('Error');
         }
     }